pyproject.toml
[tool.poetry]
name = "autogen-researcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyautogen = "^0.2.15"
jupyter = "^1.0.0"
markdown = "^3.5.2"
gitpython = "^3.1.42"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.0.2"
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
